requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
requests==2.31.0
urllib3==2.0.4
PyYAML==5.4.1